Stubbornly load images if we see MissingDelegateError
For some reason, Wand doesn't try to load ICO files as they are
if you pass in a blob and you don't specify the format.  So if
we fail to load an image, we just try loading it as an ICO file.
Detecting the format via magic numbers or something would probably
be a better long term solution